normed all functions exept for comments and long display lines restructured architecture added functions for norme (main_utils.c, exec/utils/, exec/error_handling)
