+++ Makefile
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: cdomet-d <cdomet-d@student.42lyon.fr>      +#+  +:+       +#+         #
+#    By: jauseff <jauseff@student.42lyon.fr>        +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/11/07 10:23:52 by cdomet-d          #+#    #+#              #
-#    Updated: 2024/04/05 13:02:17 by cdomet-d         ###   ########lyon.fr    #
+#    Updated: 2024/04/07 23:48:58 by jauseff          ###   ########lyon.fr    #
 #                                                                              #
 # **************************************************************************** #
 
@@ -30,7 +30,7 @@ INCLUDES := -L $(LIBFT_DIR) -lft -lreadline
 
 SRC += main.c \
 
-# ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ PARSING ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ #
+# ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ PARSING ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ #
 PARSE += $(addprefix $(DIR_TOKE), $(SRC_TOKE))
 DIR_TOKE:= parsing/tokenization/
 SRC_TOKE:=		create_data.c \
@@ -48,13 +48,13 @@ DIR_PARSE:= parsing/
 SRC_PARSE:=		create_lst.c \
 				parsing.c \
 
-# ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ LISTS ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ #
+# ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ LISTS ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ #
 SRC +=  $(addprefix $(DIR_LST), $(SRC_LST))
 DIR_LST:= lst_utils/
 SRC_LST:=		env_lst_utils.c \
 				input_lst_utils.c \
 
-# ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ BUILTINS ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ #
+# ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ BUILTINS ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ #
 SRC +=  $(addprefix $(DIR_BUILTIN), $(SRC_BUILTIN))
 DIR_BUILTIN:= builtins/
 SRC_BUILTIN:=	export.c \
@@ -62,23 +62,33 @@ SRC_BUILTIN:=	export.c \
 				exit.c \
 				builtin_utils.c \
 
-# ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ UTILS ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ #
+# ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ UTILS ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ #
 SRC +=  $(addprefix $(DIR_UTILS), $(SRC_UTILS))
 DIR_UTILS:= utils/
 SRC_UTILS:=		display_data.c \
 				display_structs.c \
-				init_input.c \
 				error_handling.c \
+				main_utils.c \
 
-# ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ EXEC ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ #
+# ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ EXEC ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ #
+
 SRC +=  $(addprefix $(DIR_EXEC), $(SRC_EXEC))
 DIR_EXEC:= exec/
 SRC_EXEC:=		exec.c \
-				exec_utils.c \
-				operators_handling.c \
 				redirections.c \
 
-# ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ RULES ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ #
+SRC +=  $(addprefix $(DIR_EXEC)$(UTILS_SUBDIR), $(SRC_SUBUTILS))
+UTILS_SUBDIR:=utils/
+SRC_SUBUTILS:=	arenvlst.c \
+				operators_handling.c \
+				redirection_utils.c \
+				fd_handling.c \
+
+SRC +=  $(addprefix $(DIR_EXEC)$(ERR_SUBDIR), $(SRC_SUBERR))
+ERR_SUBDIR:=error_handling/
+SRC_SUBERR:=	exec_errors.c \
+
+# ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ RULES ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ #
 
 OBJS:= $(addprefix $(BUILD_DIR),$(SRC:%.c=%.o))
 DEPS:= $(OBJS:%.o=%.d)
@@ -90,11 +100,11 @@ bonus : $(BONUS)
 
 $(NAME): $(LIBFT_DIR)$(LIB) $(OBJS)
 	@echo
-	@echo "$(PINK)$(BOLD) \t • • ┈┈┈┈ ๑ ⋅ ⋯ ୨ \t making minishell \t ୧ ⋯ ⋅ ๑ ┈┈┈┈ • •$(RESET)"
+	@echo "$(PB) \t • • ┈┈┈┈ ๑ ⋅ ⋯ ୨ \t making $(NAME) \t ୧ ⋯ ⋅ ๑ ┈┈┈┈ • •$(R)"
 	@echo
 	$(CC) $(CFLAGS) $(OBJS) -o $(NAME) $(INCLUDES)
 	@echo
-	@echo "$(PINK)$(BOLD) \t• • ┈┈┈┈ ๑ ⋅ ⋯ ୨ \t minishell done ! \t ୧ ⋯ ⋅ ๑ ┈┈┈┈ • •$(RESET)"
+	@echo "$(PB) \t• • ┈┈┈┈ ๑ ⋅ ⋯ ୨ \t $(NAME) done ! \t ୧ ⋯ ⋅ ๑ ┈┈┈┈ • •$(R)"
 	@echo
 	
 $(BUILD_DIR)%.o: $(SCR_DIR)%.c $(LIBFT_DIR)libft.h $(MS_H) $(EX_H) $(PA_H) Makefile
@@ -104,43 +114,42 @@ $(BUILD_DIR)%.o: $(SCR_DIR)%.c $(LIBFT_DIR)libft.h $(MS_H) $(EX_H) $(PA_H) Makef
 $(LIBFT_DIR)$(LIB): FORCE
 	@echo "$(FAINT)"
 	@$(MAKE) -C $(LIBFT_DIR)
-	@echo "$(RESET)"
+	@echo "$(R)"
 
  -include $(DEPS)
 
 clean:
 	@echo
-	@echo "$(PINK)$(BOLD)\t • • ┈┈┈┈ ๑ ⋅ ⋯ ୨ \t cleaning \t ୧ ⋯ ⋅ ๑ ┈┈┈┈ • •$(RESET)"
+	@echo "$(PB)\t • • ┈┈┈┈ ๑ ⋅ ⋯ ୨ \t cleaning \t ୧ ⋯ ⋅ ๑ ┈┈┈┈ • •$(R)"
 	@echo
 	@echo "$(FAINT)"
 	$(RM) $(BUILD_DIR)
 	make -C $(LIBFT_DIR) $@
 
-	
 fclean: clean
 	make -C $(LIBFT_DIR) $@
 	$(RM) $(NAME)
 	@echo
-	@echo "$(PINK)$(BOLD)\t • • ┈┈┈┈ ๑ ⋅ ⋯ ୨ \t done ! \t ୧ ⋯ ⋅ ๑ ┈┈┈┈ • •$(RESET)"
+	@echo "$(PB)\t • • ┈┈┈┈ ๑ ⋅ ⋯ ୨ \t done ! \t ୧ ⋯ ⋅ ๑ ┈┈┈┈ • •$(R)"
 	@echo
 
 	
 re: fclean all
 
-# ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ #
-
 V_PARAMS:= valgrind --log-file="log" --track-fds=yes --trace-children=yes --leak-check=full --suppressions=misc/supp.supp --show-leak-kinds=all
-suprun: all
+run: all
 	$(V_PARAMS) ./$(NAME)
 
+# ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ #
+
 help:
-	@echo "make \t\t $(FAINT)creates the executable$(RESET)"
-	@echo "make bonus \t $(FAINT)creates the bonus executable$(RESET)"
-	@echo "make clean \t $(FAINT)deletes object files & dependencies$(RESET)"
-	@echo "make fclean \t $(FAINT)deletes everything the makefile created$(RESET)"
-	@echo "make kitty \t $(FAINT)prints a lil kitty on the terminal$(RESET)"
-	@echo "make re \t $(FAINT)deletes all objects & dependencies and recompiles the executable$(RESET)"
-	@echo "make re_bonus \t $(FAINT)deletes all objects & dependencies and recompiles the bonus$(RESET)"
+	@echo "make \t\t $(FAINT)creates the executable$(R)"
+	@echo "make bonus \t $(FAINT)creates the bonus executable$(R)"
+	@echo "make clean \t $(FAINT)deletes object files & dependencies$(R)"
+	@echo "make fclean \t $(FAINT)deletes everything the makefile created$(R)"
+	@echo "make kitty \t $(FAINT)prints a lil kitty on the terminal$(R)"
+	@echo "make re \t $(FAINT)deletes all objects & dependencies and recompiles the executable$(R)"
+	@echo "make re_bonus \t $(FAINT)deletes all objects & dependencies and recompiles the bonus$(R)"
 
 kitty:
 	@echo "   |\__/,|   ( \  "
@@ -155,12 +164,14 @@ print% :
 .PHONY : clean fclean all re help kitty display force print%
 
 # Text
-RESET=\033[0m
+R=\033[0m
 FAINT=\033[2m
 UNDERLINE=\033[4m
 BOLD=\033[1m
 
 # Codes d'échappement ANSI pour les couleurs
+PB=$(PINK)$(BOLD)
+
 RED=\033[0;31m
 GREEN=\033[0;32m
 YELLOW=\033[0;33m
