+++ includes/parsing.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parsing.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cdomet-d <cdomet-d@student.42lyon.fr>      +#+  +:+       +#+        */
+/*   By: jauseff <jauseff@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/26 15:41:26 by csweetin          #+#    #+#             */
-/*   Updated: 2024/03/29 14:45:08 by cdomet-d         ###   ########lyon.fr   */
+/*   Updated: 2024/04/07 22:27:39 by jauseff          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,5 +25,7 @@ char	**build_tab(char *line, int *i, int word);
 int		check_quote(char *line);
 char	**get_data(t_input **input, char *line, int *i);
 void	count_word(char *line, int i, int *word);
+void	init_all(char **str, char **envp, t_env **env, t_input **input);
+void	process_line(char *line, t_input *input, t_env *env);
 
 #endif
