+++ scripts/check_log
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+# greping log to find leaks
+grep "total heap usage" log
+grep -A 500 "Open file descriptor [0-1024]:$" log | awk '1; /== $/{exit}'
