+++ src/builtins/builtin_utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   builtin_utils.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cdomet-d <cdomet-d@student.42lyon.fr>      +#+  +:+       +#+        */
+/*   By: jauseff <jauseff@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/27 17:01:23 by cdomet-d          #+#    #+#             */
-/*   Updated: 2024/04/04 17:24:05 by cdomet-d         ###   ########lyon.fr   */
+/*   Updated: 2024/04/07 22:30:26 by jauseff          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
