+++ src/exec/utils/arenvlst.c
@@ -1,0 +1,57 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   arenvlst.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jauseff <jauseff@student.42lyon.fr>        +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/03/25 14:27:46 by cdomet-d          #+#    #+#             */
+/*   Updated: 2024/04/07 23:25:31 by jauseff          ###   ########lyon.fr   */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "exec.h"
+
+static size_t	env_len(t_env *env)
+{
+	t_env	*tmp;
+	size_t	n_nodes;
+
+	tmp = env;
+	n_nodes = 0;
+	while (tmp)
+	{
+		n_nodes++;
+		tmp = tmp->next;
+	}
+	return (n_nodes);
+}
+
+char	**arenvlst(t_env	*env)
+{
+	char	**arenv;
+	size_t	i;
+	t_env	*tmp;
+
+	tmp = env;
+	if (!tmp)
+		return (NULL);
+	arenv = NULL;
+	i = 0;
+	arenv = malloc(sizeof(char *) * (env_len(env) + 1));
+	if (!arenv)
+		return (print_error(errno, NULL));
+	arenv[env_len(env)] = 0;
+	while (tmp)
+	{
+		arenv[i] = ft_strdup(tmp->env);
+		if (!arenv[i])
+		{
+			free_dtab(arenv);
+			return (print_error(errno, NULL));
+		}
+		i++;
+		tmp = tmp->next;
+	}
+	return (arenv);
+}
