+++ src/exec/utils/operators_handling.c
