+++ src/exec/utils/redirection_utils.c
@@ -1,0 +1,49 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   redirection_utils.c                                :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jauseff <jauseff@student.42lyon.fr>        +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/04/07 22:44:46 by jauseff           #+#    #+#             */
+/*   Updated: 2024/04/07 22:53:37 by jauseff          ###   ########lyon.fr   */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "exec.h"
+
+bool	is_first_cmd(t_input *in)
+{
+	t_input	*tmp;
+	bool	first;
+
+	tmp = in;
+	first = true;
+	if (tmp->prev == NULL)
+		return (first);
+	while (tmp)
+	{
+		tmp = tmp->prev;
+		if (tmp && tmp->tok == command)
+			first = false;
+	}
+	return (first);
+}
+
+bool	is_last_cmd(t_input *in)
+{
+	t_input	*tmp;
+	bool	last;
+
+	tmp = in;
+	last = true;
+	if (tmp->next == NULL)
+		return (last);
+	while (tmp)
+	{
+		tmp = tmp->next;
+		if (tmp && tmp->tok == command)
+			last = false;
+	}
+	return (last);
+}
