+++ src/utils/main_utils.c
@@ -1,33 +1,33 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   main.c                                             :+:      :+:    :+:   */
+/*   main_utils.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: jauseff <jauseff@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/03/26 15:46:56 by csweetin          #+#    #+#             */
-/*   Updated: 2024/04/07 22:27:57 by jauseff          ###   ########lyon.fr   */
+/*   Created: 2024/04/07 22:08:34 by jauseff           #+#    #+#             */
+/*   Updated: 2024/04/07 22:27:13 by jauseff          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "parsing.h"
 
-int	main(int argc, char **argv, char **envp)
+void	init_all(char **str, char **envp, t_env **env, t_input **input)
 {
-	char	*line;
-	t_env	*env;
-	t_input	*input;
+	*str = NULL;
+	*env = NULL;
+	*input = NULL;
+	create_env(input, envp, env);
+}
 
-	if (argc != 1)
-		fatal_exit(NULL, EXIT_FAILURE, "minishell: too many arguments");
-	(void)argv;
-	init_all(&line, envp, &env, &input);
-	while (1)
-	{
-		line = readline("Minishell > ");
-		if (line)
-			process_line(line, input, env);
-	}
-	env_freelst(env);
-	return (0);
+void	process_line(char *line, t_input *input, t_env *env)
+{
+	if (ft_strncmp(line, "exit", ft_strlen(line)) == 0)
+		mh_exit(line, input, env);
+	add_history(line);
+	parsing(&input, &env, line);
+	if (input)
+		exec_cmd(input);
+	input_freelst(&input);
+	free(line);
 }
