+++ test_units/debug.c
@@ -1,33 +1,28 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   main.c                                             :+:      :+:    :+:   */
+/*   debug.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: jauseff <jauseff@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/03/26 15:46:56 by csweetin          #+#    #+#             */
-/*   Updated: 2024/04/07 22:27:57 by jauseff          ###   ########lyon.fr   */
+/*   Created: 2024/04/05 12:35:14 by cdomet-d          #+#    #+#             */
+/*   Updated: 2024/04/07 22:06:53 by jauseff          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "parsing.h"
+#include "exec.h"
 
-int	main(int argc, char **argv, char **envp)
+int	main(int argc, char **argv)
 {
-	char	*line;
-	t_env	*env;
-	t_input	*input;
+	t_input	*in;
 
-	if (argc != 1)
-		fatal_exit(NULL, EXIT_FAILURE, "minishell: too many arguments");
-	(void)argv;
-	init_all(&line, envp, &env, &input);
-	while (1)
-	{
-		line = readline("Minishell > ");
-		if (line)
-			process_line(line, input, env);
-	}
-	env_freelst(env);
-	return (0);
+	in = NULL;
+	(void)argc;
+	init_input(argv, &in, NULL);
+	print_in_for(in);
+	in = in->next->next->next;
+	if (is_first_pipe(in))
+		printf("First pipe");
+	else
+		printf("Not first pipe");
 }
